+++ llmexer/exceptions.py
@@ -7,3 +7,11 @@ class LLMExerException(Exception):
 
 class ExperimentAlreadyExistsException(LLMExerException):
     """Raised when an experiment with the given ID already exists."""
+
+
+class ExperimentNotExistsException(LLMExerException):
+    """Raised when an experiment with the given ID does not exists."""
+
+
+class ExperimentIDRequiredException(LLMExerException):
+    """Raised an experiment ID has not been passed."""